refactor eval related code into module
