+++ client/construct.js
@@ -1,12 +1,13 @@
 import Editor from '../imports/editor.js';
 import ProgramHelpers from '../imports/program-helpers.js';
+import Eval from '../imports/eval.js';
 
 var Programs = new Mongo.Collection('programs');
 // https://github.com/josdirksen/learning-threejs/blob/master/chapter-09/07-first-person-camera.html
 
 var construct = null;
 
-var globalScope = window;
+
 var MODULE = ProgramHelpers.MODULE;
 
 Accounts.ui.config({
@@ -18,6 +19,7 @@ Meteor.subscribe('all-programs');
 class Construct {
   constructor($container, user) {
     console.log('initializing the construct');
+    this.evalModule = new Eval(Programs);
     this.initPhysics();
 
     this.$container = $container;
@@ -106,7 +108,7 @@ class Construct {
   initProgram(program) {
     var self = this;
     try {
-      var initializeProgram = self.evalProgramWithDependencies(
+      var initializeProgram = self.evalModule.evalProgramWithDependencies(
         program.initialize, program);
 
       var programRenderedObjects = initializeProgram(
@@ -125,71 +127,6 @@ class Construct {
       console.warn(
         `Problem initializing program ${program._id}: ${errorString}`);
     }
-  }
-
-  /*
-   Evaluate programFunction after all of the modules in the import chain/tree
-   have been evaluated (either in during this call or previously)
-   */
-  evalProgramWithDependencies(programFunction, program) {
-    var self = this;
-    var unevaluatedImports = _.reject(program.imports, hasBeenEvaluated);
-
-    _.each(unevaluatedImports, (moduleName) => {
-      self.evalModuleWithDependencies(moduleName);
-    });
-
-    return eval(programFunction);
-
-    function hasBeenEvaluated(moduleName) {
-      return _.has(globalScope, moduleName);
-    }
-  }
-
-  evalModuleWithDependencies(moduleName) {
-    var self = this;
-    var module = Programs.findOne({name: moduleName});
-    var modulesToEvaluate = module ? [module] : [];
-    // dependencies that have been evaluated in past calls to eval
-    // or have already been added to the stack of modules to evaluate
-    var modulesVisited = {};
-
-    while (!_.isEmpty(modulesToEvaluate)) {
-      var current = modulesToEvaluate.pop();
-      // if all the imports have already been evaluated
-      // evaluate current, otherwise
-      var unvisitedModuleNames = _.reject(current.imports, hasBeenVisited);
-      var unvisitedModules = _.isEmpty(unvisitedModuleNames) ? [] :
-            _.compact(Programs.find({name: unvisitedModuleNames}).fetch());
-      if (_.isEmpty(unvisitedModules)) {
-        self.evalModule(current);
-      } else {
-        modulesToEvaluate.push(current);
-
-        _.each(unvisitedModules, (unvisitedModule) => {
-          modulesToEvaluate.push(unvisitedModule);
-          modulesVisited[unvisitedModule.name] = true;
-        });
-      };
-    }
-
-    function hasBeenVisited(module) {
-      return _.has(globalScope, module.name) ||
-        _.has(modulesVisited, module.name);
-    }
-  }
-
-  evalModule(module) {
-    globalScope[module.name] = {};
-    var moduleObject = globalScope[module.name];
-    eval(module.code);
-    var parseTree = esprima.parse(module.code);
-    var declarations = _.filter(parseTree.body, (node) => {
-      return node.type.indexOf('Declaration') > 0;
-    });
-    _.each(declarations, (declaration) => {
-      moduleObject[declaration.id.name] = eval(declaration.id.name);
-    });
   }
 
   initKeyboard() {
@@ -406,6 +343,14 @@ class Construct {
     });
   }
 
+  evalTest() {
+    try {
+      eval("var a = 3");
+    } catch(e) {
+      console.log(e.message);
+    }
+  }
+
   createProgram() {
     var newProgramPosition = Programs.findOne(this.userProgramId).position;
     // the observer on the collection will render the new program
@@ -550,6 +495,7 @@ Template.hud.events({
   },
   'click .open-editor': () => {
     construct.editor.activate();
+    construct.evalTest();
   },
   'click .close-editor': () => {
     construct.editor.deactivate();
